+++ src/Report/SarifReport.php
@@ -26,7 +26,7 @@ class SarifReport implements ReportInterface
         protected readonly Environment       $twig,
         private readonly CliOutput           $output)
     {
-        $this->outputDir = $config->get('reportDir') . DIRECTORY_SEPARATOR;
+        $this->outputDir = $config->get('reportDir') . DIRECTORY_SEPARATOR . 'sarif' . DIRECTORY_SEPARATOR;
 
         if (!is_dir($this->outputDir)) {
             mkdir(directory: $this->outputDir, recursive: true);